requirements.txt
# Dependências Python do projeto.
# Incluem Flask, OpenCV, manipulação de dados e geração de gráficos.

blinker==1.9.0
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
fonttools==4.62.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.1
pandas==3.0.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
tzdata==2026.1
Werkzeug==3.1.8
